add configure code of  json to jsonp
