+++ src/main/resources/application.properties
@@ -2,3 +2,6 @@
 # Spring Boot Actuator Vulnerable Config
 management.security.enabled=false
 logging.config=classpath:logback-online.xml
+
+# jsonp check referer switch
+org.joychou.security.jsonp = true
